% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fes_taula.R
\name{fes_taula}
\alias{fes_taula}
\title{Compute a weighted crosstab.}
\usage{
fes_taula(dades, VI, VD, pes)
}
\arguments{
\item{dades}{dataset}

\item{VI}{independent variable name (as character string)}

\item{VD}{dependent variable name (as character string)}

\item{pes}{weight variable name (as character string)}
}
\value{
a tibble including counts (n), weighted counts (N), marginals of the independent variable (TT), and percentages within independent variable groups (PP = N/TT*100)
}
\description{
Compute a weighted crosstab.
}
\examples{
data(package = "palmerpenguins", "penguins")
penguins$pes <- 1
fes_taula(penguins, "island", "sex", "pes")
}
