% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{llars}
\alias{llars}
\title{Sub-mostra de joguina de la ECV 2021. Taula de llars}
\format{
Tibble.  
\describe{  
  \item{IDllar}{Identificador de la llar}  
  \item{HB030}{Pes de la llar}  
  \item{HY020}{Renda anual de la llar}  
  \item{HH010}{Tipus d'habitatge}  
  \item{HH021}{Règim de tinença}  
  \item{HS090}{Disposa la llar d'un ordinador?}  
  \item{HS120}{Capacitat de la llar d'arribar a final de mes}  
  \item{vhMATDEP}{Llar amb privació material}  
  ...  
}
}
\source{
\url{ine.es}
}
\usage{
llars
}
\description{
Sub-mostra de joguina de la ECV 2021. Taula de llars
}
\keyword{datasets}
