% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{persones}
\alias{persones}
\title{Sub-mostra de joguina de la ECV 2021. Taula de persones}
\format{
Tibble.  
\describe{  
  \item{ID}{Identificador de l'individu}  
  \item{IDllar}{Identificador de la llar}  
  \item{RB050}{Pes dels individus de la llar}  
  \item{RB090}{Sexe}  
  \item{RB080}{Any de naixement}  
  \item{roba}{Pot canviar la roba feta malbé?}  
  ...  
}
}
\source{
\url{ine.es}
}
\usage{
persones
}
\description{
Sub-mostra de joguina de la ECV 2021. Taula de persones
}
\keyword{datasets}
